refactor(tools): move gemini-list-models to gemini://models resource (#108)
The model list is static data — it belongs as an MCP Resource (cacheable
by URI, no tool catalog overhead per turn) rather than as a tool the model
must reason over on every prompt.

- Add gemini://models to STATIC_RESOURCES in src/resources.ts; readResource()
  routes to readModels() returning the same payload shape ({models, total,
  source}) as the deleted tool, with GEMINI_MODELS env override honored.
- Move the curated list + parseCustomModels into src/models-data.ts so
  resources.ts stays focused on dispatch.
- Delete src/tools/gemini-list-models.ts, tests/tools/gemini-list-models.test.ts,
  and the corresponding entries in dispatcher.ts + index.ts tool registration.
- Update tests/index.test.ts (10 tools -> 9), tests/tool-annotations.test.ts
  (6 tools -> 5), tests/resources.test.ts (3 static URIs -> 4, plus default/
  custom/URI-preservation coverage for gemini://models).
- README: add Resources section documenting gemini://models and the existing
  health/sessions/jobs URIs.

The tool surface drops from 10 to 9 tools. No deprecation alias — single-
maintainer server with no published clients depending on the tool name
(per planning decision).

640 tests pass; npm run build + npx tsc --noEmit clean.
