--- tests/tools/gemini-list-models.test.ts
@@ -1,205 +1,0 @@
-import { describe, it, expect, beforeEach, afterEach } from "vitest";
-import { ZodError } from "zod";
-import {
-  geminiListModels,
-  geminiListModelsToolDefinition,
-} from "../../src/tools/gemini-list-models.js";
-import { handleCallTool } from "../../src/dispatcher.js";
-
-describe("gemini-list-models", () => {
-  let originalModels: string | undefined;
-
-  beforeEach(() => {
-    originalModels = process.env.GEMINI_MODELS;
-    delete process.env.GEMINI_MODELS;
-  });
-
-  afterEach(() => {
-    if (originalModels !== undefined) {
-      process.env.GEMINI_MODELS = originalModels;
-    } else {
-      delete process.env.GEMINI_MODELS;
-    }
-  });
-
-  describe("default curated list", () => {
-    it("returns all default models with correct structure", async () => {
-      const result = await geminiListModels({});
-      expect(result.source).toBe("curated");
-      expect(result.total).toBe(6);
-      expect(result.models).toHaveLength(6);
-
-      for (const model of result.models) {
-        expect(model).toHaveProperty("id");
-        expect(model).toHaveProperty("description");
-        expect(model).toHaveProperty("tier");
-        expect(model).toHaveProperty("notes");
-        expect(["fast", "balanced", "deep"]).toContain(model.tier);
-      }
-    });
-
-    it("includes the expected model IDs", async () => {
-      const result = await geminiListModels({});
-      const ids = result.models.map((m) => m.id);
-      expect(ids).toContain("gemini-3-pro-preview");
-      expect(ids).toContain("gemini-3-flash-preview");
-      expect(ids).toContain("gemini-3.1-pro-preview");
-      expect(ids).toContain("gemini-2.5-pro");
-      expect(ids).toContain("gemini-2.5-flash");
-      expect(ids).toContain("gemini-2.5-flash-lite");
-    });
-
-    it("classifies gemini-2.5-flash as tier 'balanced'", async () => {
-      // Deliberate editorial reclassification from "fast" to "balanced" in the 0.38.1
-      // refresh — pin the value so an accidental revert is caught.
-      const result = await geminiListModels({});
-      const model = result.models.find((m) => m.id === "gemini-2.5-flash");
-      expect(model?.tier).toBe("balanced");
-    });
-
-    it("marks gemini-3-flash-preview with notes 'default'", async () => {
-      const result = await geminiListModels({});
-      const model = result.models.find((m) => m.id === "gemini-3-flash-preview");
-      expect(model?.notes).toBe("default");
-    });
-
-    it("marks gemini-3.1-pro-preview with notes 'limited rollout'", async () => {
-      const result = await geminiListModels({});
-      const model = result.models.find((m) => m.id === "gemini-3.1-pro-preview");
-      expect(model?.notes).toBe("limited rollout");
-    });
-  });
-
-  describe("filter", () => {
-    it("filters by substring match on model ID", async () => {
-      const result = await geminiListModels({ filter: "flash" });
-      expect(result.total).toBe(3);
-      expect(result.models.every((m) => m.id.includes("flash"))).toBe(true);
-    });
-
-    it("filters case-insensitively", async () => {
-      const result = await geminiListModels({ filter: "PRO" });
-      expect(result.total).toBe(3);
-      expect(result.models.every((m) => m.id.includes("pro"))).toBe(true);
-    });
-
-    it("returns empty array when filter matches nothing", async () => {
-      const result = await geminiListModels({ filter: "nonexistent" });
-      expect(result.models).toEqual([]);
-      expect(result.total).toBe(0);
-      expect(result.source).toBe("curated");
-    });
-  });
-
-  describe("GEMINI_MODELS env override", () => {
-    it("returns custom models when env is set", async () => {
-      process.env.GEMINI_MODELS = "custom-model-a,custom-model-b";
-      const result = await geminiListModels({});
-
-      expect(result.source).toBe("custom");
-      expect(result.total).toBe(2);
-      expect(result.models[0].id).toBe("custom-model-a");
-      expect(result.models[1].id).toBe("custom-model-b");
-      expect(result.models[0].tier).toBe("balanced");
-      expect(result.models[0].description).toBe("Custom model");
-    });
-
-    it("trims whitespace from custom model IDs", async () => {
-      process.env.GEMINI_MODELS = " model-a , model-b ";
-      const result = await geminiListModels({});
-      expect(result.models[0].id).toBe("model-a");
-      expect(result.models[1].id).toBe("model-b");
-    });
-
-    it("ignores empty segments in GEMINI_MODELS", async () => {
-      process.env.GEMINI_MODELS = "model-a,,model-b,";
-      const result = await geminiListModels({});
-      expect(result.total).toBe(2);
-    });
-
-    it("combines custom models with filter", async () => {
-      process.env.GEMINI_MODELS = "alpha-fast,beta-slow,gamma-fast";
-      const result = await geminiListModels({ filter: "fast" });
-      expect(result.total).toBe(2);
-      expect(result.source).toBe("custom");
-      expect(result.models.map((m) => m.id)).toEqual([
-        "alpha-fast",
-        "gamma-fast",
-      ]);
-    });
-  });
-
-  describe("input validation", () => {
-    it("accepts undefined input", async () => {
-      const result = await geminiListModels(undefined);
-      expect(result.total).toBe(6);
-    });
-
-    it("accepts empty object", async () => {
-      const result = await geminiListModels({});
-      expect(result.total).toBe(6);
-    });
-
-    it("rejects non-string filter", () => {
-      expect(() => geminiListModels({ filter: 123 })).toThrow(ZodError);
-    });
-  });
-
-  describe("tool definition", () => {
-    it("has correct name and title", () => {
-      expect(geminiListModelsToolDefinition.name).toBe("gemini-list-models");
-      expect(geminiListModelsToolDefinition.title).toBe("List Gemini Models");
-    });
-
-    it("has correct annotations", () => {
-      expect(geminiListModelsToolDefinition.annotations).toEqual({
-        title: "List Gemini Models",
-        readOnlyHint: true,
-        destructiveHint: false,
-        idempotentHint: true,
-        openWorldHint: false,
-      });
-    });
-
-    it("has outputSchema with expected required fields", () => {
-      expect(geminiListModelsToolDefinition.outputSchema).toBeDefined();
-      expect(
-        (geminiListModelsToolDefinition.outputSchema as { required: string[] })
-          .required
-      ).toEqual(["models", "total", "source"]);
-    });
-  });
-
-  describe("GEMINI_MODELS edge cases", () => {
-    it("returns empty custom list when GEMINI_MODELS is whitespace-only", async () => {
-      process.env.GEMINI_MODELS = "  ";
-      const result = await geminiListModels({});
-      expect(result.source).toBe("custom");
-      expect(result.total).toBe(0);
-      expect(result.models).toEqual([]);
-    });
-  });
-});
-
-describe("dispatcher routing for gemini-list-models", () => {
-  it("routes gemini-list-models through handleCallTool and returns structuredContent", async () => {
-    const result = await handleCallTool("gemini-list-models", {});
-
-    expect(result.isError).toBeUndefined();
-    expect(result.content[0].type).toBe("text");
-    const parsed = JSON.parse(result.content[0].text);
-    expect(parsed.source).toBe("curated");
-    expect(parsed.models).toBeInstanceOf(Array);
-    expect(parsed.total).toBe(6);
-    expect(result.structuredContent).toEqual(parsed);
-  });
-
-  it("passes filter through dispatcher", async () => {
-    const result = await handleCallTool("gemini-list-models", { filter: "pro" });
-
-    expect(result.isError).toBeUndefined();
-    const parsed = JSON.parse(result.content[0].text);
-    expect(parsed.total).toBe(3);
-    expect(parsed.models.every((m: { id: string }) => m.id.includes("pro"))).toBe(true);
-  });
-});
